refactor : 우승자 발견 및 문장 만들기 함수를 WinnersFinder클래스로 분할
