+++ README.md
@@ -9,9 +9,6 @@
   - [x] 전진 관리 함수
   - [x] 난수 크기 확인함수
   - [x] 우승자 출력 함수
-	- [x] 가장 많은 이동 횟수를 반환하는 함수
-	- [x] 우승자 StringBuffer를 반환하는 함수
-	- [x] 우승자 출력 시 이름 사이에 쉼표를 출력하기 위한 함수 
 
 - RaceGameManager (게임 객체를 통해 게임을 실행하고 조종하는 클래스)
   - [x] 게임 시작 함수
@@ -21,9 +18,9 @@
   - [x] 전진 함수
   - [x] 현재위치 내보내기 함수
 
-=============================
-
 - Constant (상수 집합 클래스, 게임 설정 변경시 여기서 할 수 있음)
+
+=============================
 
 - InputView (입력을 담당하는 클래스)
 
@@ -51,6 +48,11 @@
 - RepetitionValidator (시도 횟수 입력 유효성 검사를 담당하는 클래스)
   - [x] 시도 횟수가 숫자인지 여부
   - [x] 시도 횟수가 1 이상인지 여부
+
+- WinnersFinder (우승자를 찾아내는 클래스)
+  - [x] 가장 많은 이동 횟수를 반환하는 함수
+  - [x] 우승자 StringBuffer를 반환하는 함수
+  - [x] 우승자 출력 시 이름 사이에 쉼표를 출력하기 위한 함수 
 
 ## 🚀 기능 요구사항
 - 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
