+++ src/main/java/racegame/RaceGame.java
@@ -4,6 +4,7 @@ import racegame.view.InputView;
 import racegame.view.OutputView;
 import racingcar.Car;
 import utils.RandomUtils;
+import utils.WinnersFinder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -38,36 +39,8 @@ public class RaceGame {
     }
 
     public void printWinners(){
-        int maxPosition = getMaximumPosition(carList);
-        StringBuffer winners = getWinnersStringBuffer(carList, maxPosition);
-
-        OutputView.printWinners(winners.toString());
-    }
-
-    private static int getMaximumPosition(List<Car> carList){
-        List<Integer> list = new ArrayList<>();
-        for (Car car : carList) {
-            list.add(car.getPosition());
-        }
-        return list.stream().max(Integer::compare).orElse(0);
-    }
-
-    private static StringBuffer getWinnersStringBuffer(List<Car> carList, int maxPosition){
-        StringBuffer winners = new StringBuffer();
-        for(Car car : carList){
-            if(car.getPosition() == maxPosition) {
-                winners = printNameDelimiter(winners, winners.length());
-                winners.append(car.getName());
-            }
-        }
-        return winners;
-    }
-
-    private static StringBuffer printNameDelimiter(StringBuffer winners, int length){
-        if(length >0){
-            winners.append(OUTPUT_VIEW_WINNER_DELIMITER);
-        }
-        return winners;
+        String winners = WinnersFinder.getWinnersStringBuffer(carList);
+        OutputView.printWinners(winners);
     }
 
     private void manageCarMovement(){
